renamed from MacOS.h to Apple.h
